Move color functions to utility module
